+++ packages/select/src/components/ui/ChipMultiSelect.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+interface Props {}
+
+export const ChipMultiSelect: React.FC<Props> = () => {
+  return <div></div>;
+};