+++ packages/select/src/index.ts
@@ -1,4 +1,6 @@
+export * from "./components/ui/AsyncMultiSelect";
 export * from "./components/ui/AsyncSelect";
+export * from "./components/ui/MultiSelect";
 export * from "./components/ui/Select";
 export * from "./components/ui/GroupedMultiSelect";
 export * from "./SelectTheme";